mtaaAddress-backend/requirements.txt
django>=5.0
django-cors-headers>=4.7
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
